[FIX] quarknet: Broken interface fixed. Added validation for empty database. New colors added.
